+++ server/main.go
@@ -155,6 +155,7 @@ func main() {
 	mux.HandleFunc("POST /logout", app.hLogout)
 
 	mux.HandleFunc("GET /", app.requireUI(app.hUI))
+	mux.HandleFunc("GET /map", app.requireUI(app.hTopology))
 	mux.HandleFunc("GET /bladerunners/{id}", app.requireUI(app.hRackPage))
 
 	mux.HandleFunc("POST /sites", app.requireUI(app.hUISiteCreate))